pyproject.toml
[project]
name = "garmin"
version = "0.1.0"
description = "Garmin Health Dashboard"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "garminconnect>=0.2.38",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
]
